requirements-full.txt
# Web framework
fastapi==0.129.2
uvicorn[standard]==0.40.0
sse-starlette==3.1.2
python-multipart==0.0.21

# Claude Agent SDK
anthropic==0.75.0
claude-agent-sdk==0.1.18

# Async support
anyio==4.12.1
httpx==0.28.1
httpx-sse==0.4.3
sniffio==1.3.1

# Templating
Jinja2==3.1.6
MarkupSafe==3.0.3
Markdown==3.7

# Utils
python-dotenv==1.2.1
click==8.3.1
requests==2.32.3
PyYAML==6.0.2

# Pydantic for SDK
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.12.0

# Type hints
typing_extensions==4.15.0

# Production server (uvicorn already listed above)

# Database
psycopg2-binary==2.9.10

# S3-compatible object storage
boto3==1.35.0
pexpect==4.9.0

# Data analysis
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.4.0
hmmlearn>=0.3.0

# Embeddings (research corpus)
sentence-transformers>=3.0.0

# File format support
openpyxl>=3.1.0      # Excel .xlsx
mammoth>=1.6.0       # Word .docx → markdown
pdfplumber>=0.10.0   # PDF text extraction
